sentiment_service/requirements.txt
tensorflow>=2.16.1
transformers>=4.37.0
torch>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
flask>=3.0.0
python-dotenv>=1.0.0
flask-cors>=4.0.0
